Add morphexsynth: an MPE subtractive synth example ported from modabacad's NubuNiif
Full engine (oscillators, resonant filter, envelopes, LFO, distortion, a
10-slot MPE routing matrix), a 12-voice MPE-zone-aware voice pool, and Lua
scripting, all with unit tests. Incidentally fixes a handful of pre-existing
compiler warnings in Wavetables/SynthHandling headers this port was the
first to compile under the full project's stricter warning set, and a
generator bug in parseboxstructure.py's single-column layout case.

+++ src/includes/SynthHandling/SustainPedalHandler.h
@@ -98,7 +98,7 @@ class SustainPedalHandler
             if (slotIndex == -1)
             {
                 slotIndex = findVoiceToSteal();
-                auto& stolenSlot = m_activeNotes[slotIndex];
+                auto& stolenSlot = m_activeNotes[static_cast<size_t>(slotIndex)];
                 if (m_noteOff)
                 {
                     m_noteOff(stolenSlot.channel, stolenSlot.note, stolenSlot.noteOffVelocity);
@@ -107,7 +107,7 @@ class SustainPedalHandler
             }
         }
 
-        auto& slot = m_activeNotes[slotIndex];
+        auto& slot = m_activeNotes[static_cast<size_t>(slotIndex)];
         slot = {true, false, channel, note, velocity, m_counter++};
 
         // Always trigger noteOn (MIDI spec requirement)
